Move constraint tests to feature tests (#22)
* mv constraint to feature test

* sort-imports

* better msg

* add json validation check
